requirements.txt
langchain==1.2.6
langchain-core==1.2.7
langchain-openai==1.1.7
langfuse==4.2.0
python-dotenv==1.2.2
